tests/authServerSpec.js - add a test for a base64 enconded nonce (includes = chars)
This nonce, taken from a real-world server (Restlet 2.3) fails,
because the server does not get back a nonce string of the expected
length (64, the length it sent out).  This is because the
parseHeader() function does not recognise the '=' character, and
simply omits it from the parsed value.

More generally, the parser can fail to parse silently, when characters
are included in quotes but not included in the _valuePattern regex's
allowed characters.